refactor: move polytope topology + generators to rye-shape, decouple rye-render from rye-physics
Polytope4 topology (polytope.rs) and the convex-polytope vertex/face-plane generators + Wolfe SDF (euclidean_r4 449-831, pure glam) move from rye-physics to rye-shape (the dependency-free base crate), as new modules polytope + polytope_geom. rye-physics re-exports both (its collision code uses them, so this is a real re-export, not a compat shim) so euclidean_r4 sim code and existing rye_physics::{polytope,euclidean_r4} consumers keep working. rye-render now imports Polytope4 + generators from rye-shape and DROPS its rye-physics dependency entirely, fixing the render->physics layering smell. Workspace builds, clippy -D warnings clean, 771 tests pass, docs clean.
