+++ crates/rye-render/Cargo.toml
@@ -19,7 +19,6 @@ winit.workspace = true
 glam.workspace = true
 bytemuck.workspace = true
 rye-math    = { path = "../rye-math" }
-rye-physics = { path = "../rye-physics" }
 rye-shape   = { path = "../rye-shape" }
 # Frame-trace integration: `gpu_timer` callbacks push `gpu-total` sections through
 # `rye_time::frame_trace::record_external` so they appear in `trace summary` next
